organize the redux folder structure
